Updated SPSA Tuner

+++ STiltedBase.cpp
@@ -347,5 +347,5 @@ void TTentry::print(){
 
 struct EvalVars{
     double* rc;
-    int* aw;
+    double* aw;
 };
